requirements.txt
pyobjc==10.2
PySide6==6.7.0